+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -94,5 +94,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-10T14:55:44.900Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-11T11:09:43.934Z")
     r"""The date and time the resource was created"""