+++ packages/orq-rc/RELEASES.md
@@ -759,3 +759,13 @@ Based on:
 - [python v3.2.0-rc.38] packages/orq-rc
 ### Releases
 - [PyPI v3.2.0-rc.38] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.38 - packages/orq-rc
+
+## 2025-03-11 11:10:56
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.514.1 (2.546.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.2.0-rc.39] packages/orq-rc
+### Releases
+- [PyPI v3.2.0-rc.39] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.39 - packages/orq-rc